Refs #34322 - move ErrataTab
