+++ webpack/components/extensions/HostDetails/Tabs/ErrataTab/ErrataTab.js
@@ -20,19 +20,19 @@ import { selectAPIResponse } from 'foremanReact/redux/API/APISelectors';
 import IsoDate from 'foremanReact/components/common/dates/IsoDate';
 import { urlBuilder } from 'foremanReact/common/urlHelpers';
 import { propsToCamelCase } from 'foremanReact/common/helpers';
-import SelectableDropdown from '../../../SelectableDropdown';
-import { useSet, useBulkSelect, useUrlParams } from '../../../../components/Table/TableHooks';
-import TableWrapper from '../../../../components/Table/TableWrapper';
-import { ErrataType, ErrataSeverity } from '../../../../components/Errata';
-import { getInstallableErrata, regenerateApplicability, applyViaKatelloAgent } from '../HostErrata/HostErrataActions';
+import SelectableDropdown from '../../../../SelectableDropdown';
+import { useSet, useBulkSelect, useUrlParams } from '../../../../../components/Table/TableHooks';
+import TableWrapper from '../../../../../components/Table/TableWrapper';
+import { ErrataType, ErrataSeverity } from '../../../../../components/Errata';
+import { getInstallableErrata, regenerateApplicability, applyViaKatelloAgent } from './HostErrataActions';
 import ErratumExpansionDetail from './ErratumExpansionDetail';
 import ErratumExpansionContents from './ErratumExpansionContents';
-import { selectHostErrataStatus } from '../HostErrata/HostErrataSelectors';
-import { HOST_ERRATA_KEY, ERRATA_TYPES, ERRATA_SEVERITIES, TYPES_TO_PARAM, SEVERITIES_TO_PARAM, PARAM_TO_FRIENDLY_NAME } from '../HostErrata/HostErrataConstants';
-import { installErrata } from './RemoteExecutionActions';
-import { errataInstallUrl } from './customizedRexUrlHelpers';
+import { selectHostErrataStatus } from './HostErrataSelectors';
+import { HOST_ERRATA_KEY, ERRATA_TYPES, ERRATA_SEVERITIES, TYPES_TO_PARAM, SEVERITIES_TO_PARAM, PARAM_TO_FRIENDLY_NAME } from './HostErrataConstants';
+import { installErrata } from '../RemoteExecutionActions';
+import { errataInstallUrl } from '../customizedRexUrlHelpers';
 import './ErrataTab.scss';
-import hostIdNotReady from '../HostDetailsActions';
+import hostIdNotReady from '../../HostDetailsActions';
 
 export const ErrataTab = () => {
   const hostDetails = useSelector(state => selectAPIResponse(state, 'HOST_DETAILS'));
