+++ webpack/components/extensions/HostDetails/Tabs/ErrataTab/ErrataTab.scss
