+++ webpack/components/extensions/HostDetails/Tabs/ErrataTab/ErratumExpansionContents.js
