+++ webpack/components/extensions/HostDetails/Tabs/ErrataTab/ErratumExpansionDetail.js
