+++ webpack/components/extensions/HostDetails/Tabs/ErrataTab/HostErrataConstants.js
